Extend Layer::Mask to support up to 128 layers.
* Change: In Hurricane::Mask, modify the template so it can
    handle boost::multiprecision::uint128_t some operator
    overload and constructors needed to be adated to this
    new king of type.
* Change: In Layer::Mask, build on uint128_t bit to support
    128 BasicLayers.

+++ hurricane/src/hurricane/hurricane/Query.h
@@ -312,7 +312,7 @@ namespace Hurricane {
 
   class Query {
     public:
-      typedef Hurricane::Mask<int> Mask;
+      typedef Hurricane::Mask<uint64_t> Mask;
     public:
     // Types.
       enum  QueryFilter { DoMasterCells   =  1
